requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==25.12.0
bleach==6.3.0
blinker==1.9.0
catboost==1.2.10
category_encoders==2.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.1
filelock==3.25.0
Flask==3.1.3
fonttools==4.60.1
fsspec==2026.2.0
graphviz==0.21
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
imageio==2.37.0
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
lazy_loader==0.4
librt==0.7.7
lightgbm==4.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.18.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==3.0.1
pathspec==1.0.3
patsy==1.0.2
pillow==12.0.0
platformdirs==4.5.1
plotly==6.6.0
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
pyglet==2.1.10
Pygments==2.19.2
PyMuPDF==1.27.1
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-multipart==0.0.22
python-slugify==8.0.4
pytokens==0.3.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
starlette==0.52.1
statsmodels==0.14.6
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==5.3.0
trimesh==4.9.0
typer==0.24.1
types-requests==2.32.4.20260107
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
webencodings==0.5.1
Werkzeug==3.1.6
xgboost==3.2.0